+++ projects/ngx-skeleton-loader/src/lib/ngx-skeleton-loader.component.spec.ts
@@ -7,9 +7,17 @@ import { NgxSkeletonLoaderComponent } from './ngx-skeleton-loader.component';
   selector: 'ngx-container',
   template: `
     <div>
-      <div class="skeletons-defaults"><ngx-skeleton-loader></ngx-skeleton-loader></div>
+      <div class="skeletons-defaults">
+        <ngx-skeleton-loader></ngx-skeleton-loader>
+      </div>
 
-      <div class="skeletons-with-count"><ngx-skeleton-loader count="2"></ngx-skeleton-loader></div>
+      <div class="skeletons-appearance-no-animation">
+        <ngx-skeleton-loader [animation]="false"></ngx-skeleton-loader>
+      </div>
+
+      <div class="skeletons-with-count">
+        <ngx-skeleton-loader count="2"></ngx-skeleton-loader>
+      </div>
 
       <div class="skeletons-appearance-circle">
         <ngx-skeleton-loader appearance="circle"> </ngx-skeleton-loader>
@@ -25,7 +33,7 @@ import { NgxSkeletonLoaderComponent } from './ngx-skeleton-loader.component';
     </div>
   `,
 })
-class ContainerComponent { }
+class ContainerComponent {}
 
 describe('NgxSkeletonLoaderComponent', () => {
   let fixture: any;
@@ -40,23 +48,38 @@ describe('NgxSkeletonLoaderComponent', () => {
 
   describe('When skeleton is created using default settings', () => {
     it('should render a single skeleton', () => {
-      expect(fixture.nativeElement.querySelectorAll('.skeletons-defaults .loader').length).toBe(1);
+      expect(
+        fixture.nativeElement.querySelectorAll('.skeletons-defaults .loader')
+          .length
+      ).toBe(1);
     });
   });
 
   describe('When skeleton is created with count', () => {
     it('should render skeleton based on given count attribute', () => {
-      expect(fixture.nativeElement.querySelectorAll('.skeletons-with-count .loader').length).toBe(
-        2,
-      );
+      expect(
+        fixture.nativeElement.querySelectorAll('.skeletons-with-count .loader')
+          .length
+      ).toBe(2);
     });
   });
 
   describe('When skeleton is created with circle appearance', () => {
     it('should add styles based on circle class on the skeleton components', () => {
       expect(
-        fixture.nativeElement.querySelectorAll('.skeletons-appearance-circle .loader.circle')
-          .length,
+        fixture.nativeElement.querySelectorAll(
+          '.skeletons-appearance-circle .loader.circle'
+        ).length
+      ).toBe(1);
+    });
+  });
+
+  describe('When skeleton is created without animation', () => {
+    it('should NOT add progress animation styles based on animation class on the skeleton components', () => {
+      expect(
+        fixture.nativeElement.querySelectorAll(
+          '.skeletons-appearance-no-animation .loader:not(.animation)'
+        ).length
       ).toBe(1);
     });
   });
@@ -64,11 +87,11 @@ describe('NgxSkeletonLoaderComponent', () => {
   describe('When skeleton is created with theming', () => {
     it('should render skeleton with styles based on theme attribute', () => {
       const skeletonWithTheming = fixture.nativeElement.querySelector(
-        '.skeletons-with-theming .loader.circle',
+        '.skeletons-with-theming .loader.circle'
       ).attributes as NamedNodeMap;
 
       expect(skeletonWithTheming.getNamedItem('style').value).toBe(
-        'width: 70px; height: 70px; border-radius: 10px;',
+        'width: 70px; height: 70px; border-radius: 10px;'
       );
     });
   });